# mudeopensrvapi/src/java/it/csi/mudeopen/mudeopensrvapi/business/be/web/ScaricoXmlApi.java
/*
 *  SPDX-FileCopyrightText: Copyright 2020 - 2021 | CSI Piemonte
 *  SPDX-License-Identifier: EUPL-1.2
 */
package it.csi.mudeopen.mudeopensrvapi.business.be.web;

import javax.ws.rs.Consumes;
import javax.ws.rs.GET;
import javax.ws.rs.HeaderParam;
import javax.ws.rs.Path;
import javax.ws.rs.PathParam;
import javax.ws.rs.Produces;
import javax.ws.rs.core.MediaType;
import javax.ws.rs.core.Response;

@Path("")
@Consumes({MediaType.APPLICATION_JSON})
@Produces({MediaType.APPLICATION_JSON})
public interface ScaricoXmlApi {

	
	@GET
    @Path("/scarico-xml/{numeroIstanza}/{codiceTipoTracciato}")
    Response scaricoXML(@HeaderParam("X-Request-ID") String XRequestId, @HeaderParam("X-Forwarded-For") String XForwardedFor, @HeaderParam("fruitore") String fruitoreID,
    						@PathParam("numeroIstanza") String numeroIstanza, 
    						@PathParam("codiceTipoTracciato") String codiceTipoTracciato) throws Exception;
}