# mudeopensrvapi/src/yaml/mudeopenapiws.yaml
...
swagger: '2.0'
info:
  title: MUDEOPEN WS-API 10-01-2023
  description: API per i WS REST
  version: 1.0.0
host: 'localhost:8080'
schemes:
  - http
basePath: /mudeopen/ws/rest
definitions:
...
  DatiSintesiIstanzaVO:
    type: object
    properties:
      numeroIstanza:
        type: string
      tipoIstanza:
        type: string
      speciePratica:
        type: string
      tipoIntestatario:
        type: string
      cognome:
        type: string
      nome:
        type: string
      ragioneSociale:
        type: string
      professionista:
        type: string
      sedime:
        type: string
      descrizione:
        type: string
      numeroCivico:
        type: string
      bis:
        type: string
      interno:
        type: string
      comune:
        type: string
      statoIstanza:
        type: string
      istanzaDiRiferimento:
        type: string
      numeroIntervento:
        type: string
  DatiXmlVO:
    properties:
      tracciatoXML:
        type: string
      versione:
        type: string
      descrizioneTipoTracciato:
        type: string
    type: object
...
  DocumentoPraticaVO:
    properties:
      codiceTipologia:
        type: string
      tipologia:
        type: string
      nomeFile:
        type: string
      dataCaricamento:
        type: string
      notificato:
        type: boolean
      identificativo:
        format: int64
        type: integer
    type: object 
  StatiIstanzaAmmessiVO:
    properties:
      codice:
        type: string
      descrizione:
        type: string
      oggettoProposto:
        type: string
      messaggioProposto:
        type: string
    type: object
  StatoIstanzaVO:
    properties:
      codice:
        type: string
      descrizione:
        type: string
    type: object 
...
  DettaglioPraticaVO:
    type: object
    properties:
      numeroPratica:
        type: string
      comune:
        type: string
      anno:
        type: integer
        format: int32
  DefaultResponse:
    properties:
      code:
        type: string
      detail:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
  ErrorResponse:
    properties:
      status:
        type: string
      code:
        type: string
      title:
        type: string
      detail:
        type: object
        additionalProperties:
          type: string
      links:
        type: array
        items:
          type: string
    type: object
#
# 2CHECK
#
...
  DatiModificaStatoIstanza:
    properties:
      numeroIstanza:
        type: string
      nuovoStato:
        type: string
      oggetto:
        type: string
      testo:
        type: string
      numeroPraticaComunale:
        type: string
      annoPratica:
        type: integer
      numeroProtocollo:
        type: string
      dataProtocollo:
        type: string
      responsabileProcedimento:
        type: string
    type: object
...
paths:
...
  /scarico-xml/{numeroIstanza}/{codiceTipoTracciato}:
    get:
      operationId: scaricoXML
      produces:
        - application/octet-stream
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          required: true
          type: string
        - in: path
          name: codiceTipoTracciato
          required: true
          type: string
      responses:
        '200':
          description: successful response
          schema:
            $ref: '#/definitions/DatiXmlVO'
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /stato-istanza:
    put:
      operationId: modificaStatoIstanza
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: body
          name: statoIstanza
          schema:
            $ref: '#/definitions/DatiModificaStatoIstanza'
      responses:
        '200':
          description: successful response
          schema:
            type: boolean
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Index not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /istanza/{numeroIstanza}:
    get:
      operationId: ricercaDatiSintesiIstanza
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: numeroIstanza
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: successful response
          schema:
            $ref: '#/definitions/DatiSintesiIstanzaVO'
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Index not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
...
  /elenco-documenti-pratica/{istatComune}:
    get:
      operationId: ricercaElencoDocumentiPratica
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: istatComune
          type: string
          required: true
        - in: query
          name: numeroPratica
          type: string
          required: true
        - in: query
          name: annoPratica
          type: integer
          required: true
      responses:
        '200':
          description: successful operation
          schema:
            items:
              $ref: '#/definitions/DocumentoPraticaVO'
            type: array
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
...
  /stati-istanza-ammessi/{statoIniziale}:
    get:
      operationId: ricercaStatiIstanzaAmmessi
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: statoIniziale
          type: string
          required: true
      responses:
        '200':
          description: successful response
          schema:
            items:
              $ref: '#/definitions/StatiIstanzaAmmessiVO'
            type: array
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /stati-istanza:
    get:
      operationId: ricercaStatiIstanza
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
      responses:
        '200':
          description: successful response
          schema:
            items:
              $ref: '#/definitions/StatoIstanzaVO'
            type: array
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
...
  /documento-pratica/{istatComune}/{numeroPratica}/{anno}:
    get:
      operationId: estraiDocumentoPratica
      produces:
        - application/octet-stream
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: istatComune
          required: true
          type: string
        - in: path
          name: numeroPratica
          required: true
          type: string
        - in: path
          name: anno
          required: true
          type: integer
      responses:
        '200':
          description: successful response
          schema:
            type: file
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
  /elimina-documento-pratica/{istatComune}/{numeroPratica}/{anno}:
    delete:
      operationId: eliminaDocumentoPratica
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: istatComune
          required: true
          type: string
        - in: path
          name: numeroPratica
          required: true
          type: string
        - in: path
          name: anno
          required: true
          type: integer
        - in: query
          name: nomeDocumento
          type: string
          required: true
      responses:
        '200':
          description: successful response
          schema:
            type: boolean
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /pratiche/{istatComune}:
    get:
      operationId: ricercaPratiche
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: path
          name: istatComune
          type: string
          required: true
        - in: query
          name: numeroPratica
          type: string
        - in: query
          name: annoPratica
          type: integer
      responses:
        '200':
          headers:
            X-Page:
              type: integer
              format: int32
              description: For pagination purposes. The current page.
            X-Page-Size:
              type: integer
              format: int32
              description: For pagination purposes. The page size.
            X-Total-Elements:
              type: integer
              format: int32
              description: For pagination purposes. The total number of elements satisfying the search query.
            X-Total-Pages:
              type: integer
              format: int32
              description: For pagination purposes. The total number of pages for representing the total number of elements satisfying the query.
          description: successful operation
          schema:
            items:
              $ref: '#/definitions/DettaglioPraticaVO'
            type: array
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Index not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
...
  /upload-documento-pratica:
    post:
      operationId: allegaDocumentoPratica
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Request-ID
          type: string
        - in: header
          name: X-Forwarded-For
          type: string
        - in: header
          name: fruitore
          type: string
        - in: formData
          name: istatComune
          type: string
          required: true
        - in: formData
          name: numeroPratica
          type: string
          required: true
        - in: formData
          name: anno
          type: integer
          required: true
        - in: formData
          name: tipoDocumento
          type: string
          required: true
        - in: formData
          name: nomeDocumento
          type: string
          required: true
        - in: formData
          name: file
          required: true
          type: file
      responses:
        '201':
          description: successful response
          schema:
            $ref: '#/definitions/DocumentoPraticaVO'
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/DefaultResponse'
        '404':
          description: Not found
        '401':
          description: Authorization is missing or not valid
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorResponse'
...